remarcable/requirements.txt
Django==5.2.3,
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0